Refactor the way how socket io is initialized. Get socket io script link from login response and initialize it from there
