+++ package.json
@@ -11,11 +11,7 @@
     "start": "npm run docs"
   },
   "dependencies": {
-    "core-js": "^2.6.5",
-    "socket.io-client": "1.3.7"
-  },
-  "peerDependencies": {
-    "socket.io-client": "1.3.7"
+    "core-js": "^2.6.5"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "^3.8.0",
